questllama/requirements.txt
ipykernel
sentence-transformers
tqdm
llama-cpp-python
langchainhub
chromadb
gpt4all